Create csvreader.py
A Python csv reader I made for practice. It reads in the data from the csv file and prints it all out. If you want to search for sales numbers of a specific color and their respective dates, the program will find all instances of that color and append the indices to a list and then iterate through that list to the color,sales quantity and date lists according to their respective indices.

Made using this video: https://www.youtube.com/watch?v=K_oXb04izZM
and big help from the answer on this: http://stackoverflow.com/questions/6294179/how-to-find-all-occurrences-of-an-element-in-a-list