Create copyto.py
A program that will read the contents of one file check the size. Copy the contents and write it to a new file. If the file does not exist then the program will create the file before writing... next program i'll have to take it up a notch and create my own program to do something. Gotta get better at this. Since i'll probably be doing this when i get my first job........maybe not